+++ cmd/cmd.go
@@ -10,6 +10,7 @@ import (
 	"github.com/alwqx/sec/cmd/balancesheet"
 	"github.com/alwqx/sec/cmd/bond"
 	"github.com/alwqx/sec/cmd/insider"
+	"github.com/alwqx/sec/cmd/ipo"
 	"github.com/alwqx/sec/cmd/kline"
 	"github.com/alwqx/sec/cmd/metal"
 	"github.com/alwqx/sec/cmd/quote"
@@ -81,6 +82,7 @@ func NewCLI() *cobra.Command {
 		watch.NewWatchCLI(),
 		announcements.NewAnnouncementsCLI(),
 		insider.NewInsiderCLI(),
+		ipo.NewIPOCLI(),
 	)
 
 	return rootCmd